+++ lib/src/result_builder/widget_builders.dart
@@ -2,8 +2,7 @@ import 'package:flutter/widgets.dart';
 import 'package:stated_result/stated_result.dart';
 
 /// Widget builder that accepts an error/exception with its optional stack trace
-typedef Widget FailedResultBuilder(
-    BuildContext context, ErrorWithStack errorInfo);
+typedef Widget FailedResultBuilder(BuildContext context, ErrorResult errorInfo);
 
 /// Widget builder that accepts a given value but without child
 typedef Widget ValueResultBuilder<T>(BuildContext context, T value);